fingerprint-enhancer 2/requirements.txt
opencv-python-headless>=4.10.0.84
numpy>=1.26.4
matplotlib>=3.9.0
Pillow>=10.4.0
scipy>=1.13.1
scikit-image>=0.24.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.30.0
loguru>=0.7.2
typer>=0.12.3
